ci: update workflow-template
Fetch and reset to the specified branch during CI builds.

Previously, the workflow was fetching and resetting to HEAD
of the default branch. This change updates the workflow to
fetch and reset to the branch specified by the
`app.github.branch` value.  This allows for building and
deploying from different branches.  The initial clone also
now clones the specified branch.